genny no longer hang after a few minutes

+++ jsmooch-lib/src/component/audio/ym2612/ym2612.h
@@ -38,7 +38,7 @@ struct ym2612 {
     } io;
 
     struct {
-        i32 busy_until;
+        u64 busy_until;
         u64 env_cycle_counter;
     } status;
